Make tests resilient to missing matplotlib
