+++ tests/conftest.py
@@ -1,3 +1,14 @@
-import matplotlib
+from pathlib import Path
+import sys
 
-matplotlib.use('Agg')
+ROOT = Path(__file__).resolve().parents[1]
+SRC_PATH = ROOT / "src"
+if str(SRC_PATH) not in sys.path:
+    sys.path.insert(0, str(SRC_PATH))
+
+try:
+    import matplotlib
+except ModuleNotFoundError:
+    matplotlib = None
+else:
+    matplotlib.use("Agg")
